+++ docs/src/changelog.md
@@ -8,6 +8,9 @@ See the full [CHANGELOG.md](https://github.com/bug-ops/zeph/blob/main/CHANGELOG.
 
 ## [Unreleased]
 
+### Added
+- Orchestrator provider option in `zeph init` wizard for multi-model routing setup (#597)
+
 ## [0.11.0] - 2026-02-19
 
 ### Added